+++ src/debug_functions.php
@@ -43,24 +43,27 @@ if (!function_exists('debug') && function_exists('dump')) {
      * Prints out debug information about given variable.
      *
      * Alias for the `dump()` global function provided by `VarDumper` component.
+     * @param mixed $var Variable you want to debug
      * @return void
+     * @link https://symfony.com/doc/current/components/var_dumper.html#the-dump-function
      * @since 1.2.11
      */
-    function debug()
+    function debug($var)
     {
-        call_user_func_array('dump', func_get_args());
+        call_user_func('dump', $var);
     }
 }
 
 if (!function_exists('dd') && function_exists('dump')) {
     /**
      * Prints out debug information about given variable and dies
+     * @param mixed $var Variable you want to debug
      * @return void
      * @since 1.2.11
      */
-    function dd()
+    function dd($var)
     {
-        call_user_func_array('debug', func_get_args());
+        call_user_func('dump', $var);
         die(1);
     }
 }