+++ src/filesystem_functions.php
@@ -64,7 +64,7 @@ if (!function_exists('create_tmp_file')) {
      *  a stream resource
      * @param string|null $dir The directory where the temporary filename will
      *  be created
-     * @param string|null $prefix The prefix of the generated temporary filename
+     * @param string $prefix The prefix of the generated temporary filename
      * @return string Path of temporary filename
      * @deprecated Use instead `Filesystem::addSlashTerm()`
      * @since 1.1.7