+++ src/global_functions.php
@@ -115,7 +115,7 @@ if (!function_exists('is_json')) {
 if (!function_exists('is_positive')) {
     /**
      * Checks if a string is a positive number
-     * @param string|int $string String
+     * @param float|string|int $string String
      * @return bool
      */
     function is_positive($string)
@@ -133,7 +133,7 @@ if (!function_exists('is_stringable')) {
      */
     function is_stringable($var)
     {
-        return method_exists($var, '__toString') || (is_scalar($var) && !is_null($var));
+        return method_exists($var, '__toString') || is_scalar($var);
     }
 }